requirements.txt
playwright==1.44.0
tenacity==8.4.1
Jinja2==3.1.4
pandas==2.2.2
PyYAML==6.0.1
